Converted number variables and conversions to double to allow longer decimal numeral return value.

+++ Calculator/Program.cs
@@ -1,7 +1,7 @@
 ﻿// See https://aka.ms/new-console-template for more information
 
 // Declare variables and then initialise to zero
-float num1 = 0; float num2 = 0;
+double num1 = 0; double num2 = 0;
 
 // Display title as the C# console calculator app.
 Console.WriteLine("Console Calculator in C#\r");
@@ -9,11 +9,11 @@ Console.WriteLine("------------------------\n");
 
 // Ask the user to type the first number.
 Console.WriteLine("Type a number, and then press Enter");
-num1 = Convert.ToInt32(Console.ReadLine());
+num1 = Convert.ToDouble(Console.ReadLine());
 
 // Ask the user to type the second number.
 Console.WriteLine("Type another number, and then press Enter");
-num2 = Convert.ToInt32(Console.ReadLine());
+num2 = Convert.ToDouble(Console.ReadLine());
 
 // Ask the user to choose and option.
 Console.WriteLine("Choose and option from the following list:");
